power: oplus_battery_msm8250: Unregister the typec partner only during cable removal
When port-partner is unregistered we observe a drp request from
USB HAL which introduces a delay in updating the switched-mode.
Fix this by unregistering the partner only during cable removal.

While at it, use try.snk in configuration of typec without
which it is stuck in sink only mode preventing a role-switch.

Change-Id: Ic70051762e4bd6de9df8bb0c11bc2214ee15af5b
Signed-off-by: Kavya Nunna <knunna@codeaurora.org>